+++ packages/@ionic/cli-utils/src/definitions.ts
@@ -153,6 +153,11 @@ export interface ProjectFile {
   readonly integrations: ProjectIntegrations;
   readonly hooks: Record<HookName, string | string[] | undefined>;
 
+  ssl?: {
+    key?: string;
+    cert?: string;
+  };
+
   /**
    * @deprecated
    */
@@ -471,6 +476,7 @@ export interface ServeOptions {
   // Command Options
   address: string;
   port: number;
+  ssl: boolean;
   livereload: boolean;
   proxy: boolean;
   lab: boolean;