+++ packages/@ionic/cli-utils/src/lib/project/angular/serve.ts
@@ -32,6 +32,8 @@ export class ServeRunner extends BaseServeRunner<AngularServeOptions> {
       longDescription: `
 ${chalk.green('ionic serve')} uses the Angular CLI. Common Angular CLI options such as ${chalk.green('--target')} and ${chalk.green('--environment')} are mixed in with Ionic CLI options. Use ${chalk.green('ng serve --help')} to list all options. See the ${chalk.green('ng build')} docs${chalk.cyan('[1]')} for explanations. Options not listed below are considered advanced and can be passed to the Angular CLI using the ${chalk.green('--')} separator after the Ionic CLI arguments. See the examples.
 
+For serving your app with HTTPS, use the ${chalk.green('--ssl')} option. You can provide your own SSL key and certificate with the ${chalk.green('ionic config set ssl.key <path>')} and ${chalk.green('ionic config set ssl.cert <path>')} commands. If these config entries are not set, a self-signed certificate is generated.
+
 If a ${chalk.bold('proxy.config.json')} or ${chalk.bold('proxy.config.js')} file is detected in your project, the Angular CLI's ${chalk.green('--proxy-config')} option is automatically specified. You can use ${chalk.green('--no-proxy')} to disable this behavior. See the Angular CLI proxy documentation${chalk.cyan('[2]')} for more information.
 
 ${chalk.cyan('[1]')}: ${chalk.bold('https://github.com/angular/angular-cli/wiki/build#ng-build')}
@@ -63,6 +65,13 @@ ${chalk.cyan('[2]')}: ${chalk.bold('https://github.com/angular/angular-cli/wiki/
           groups: [OptionGroup.Advanced],
           hint: 'ng',
         },
+        {
+          name: 'ssl',
+          description: 'Use HTTPS for the dev server',
+          aliases: ['s'],
+          type: Boolean,
+          hint: 'ng',
+        },
       ],
     };
   }
@@ -82,6 +91,7 @@ ${chalk.cyan('[2]')}: ${chalk.bold('https://github.com/angular/angular-cli/wiki/
 
     return {
       ...baseOptions,
+      ssl: options['ssl'] ? true : false,
       target,
       environment,
     };
@@ -102,7 +112,7 @@ ${chalk.cyan('[2]')}: ${chalk.bold('https://github.com/angular/angular-cli/wiki/
 
     return {
       custom: program !== DEFAULT_PROGRAM,
-      protocol: 'http',
+      protocol: options.ssl ? 'https' : 'http',
       localAddress: 'localhost',
       externalAddress: externalIP,
       externalNetworkInterfaces: availableInterfaces,
@@ -206,10 +216,27 @@ ${chalk.cyan('[2]')}: ${chalk.bold('https://github.com/angular/angular-cli/wiki/
       progress: 'false',
       target: options.target,
       environment: options.environment,
+      ssl: options.ssl ? 'true' : undefined,
 
       // Added so Cordova doesn't complain about www directory missing.
       deleteOutputPath: options.engine === 'cordova' ? 'false' : undefined,
     };
+
+    const project = await this.env.project.load();
+
+    if (project.ssl) {
+      if (project.ssl.key && project.ssl.cert) {
+        args['ssl-key'] = path.resolve(this.env.project.directory, project.ssl.key);
+        args['ssl-cert'] = path.resolve(this.env.project.directory, project.ssl.cert);
+      } else {
+        this.env.log.warn(
+          `Both ${chalk.green('ssl.key')} and ${chalk.green('ssl.cert')} config entries must be set.\n` +
+          `See ${chalk.green('ionic serve --help')} for details on using your own SSL key and certificate for the dev server. Using self-signed certificate.`
+        );
+
+        args.ssl = true;
+      }
+    }
 
     if (options.proxy) {
       const proxyConfig = await this.detectProxyConfig();